[TE-4941] Add pytest-xdist example
A few customers use pytest-xdist with the Python test collector and
it would be nice for us to have something to demonstrate this
functionality given the issues we've had with supporting xdist
in the past.

This will also help us diagnose and replicate any future issues
customer may report with xdist related issues with the Python
test collector.

I've tagged the executions with 'test.framework.xdist=true' for
the tests executed in the pytest-xdist directory. The existing
pytest directory tests have their executions tagged with
'test.framework.xdist=false'.